% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{log.likelihood}
\alias{log.likelihood}
\title{Log-likelihood Function}
\usage{
\method{log}{likelihood}(x, pi, mu, sigma, nu)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{pi}{A numeric vector of the mixture weights \eqn{\pi_k}.}

\item{mu}{A numeric vector of the location parameter \eqn{\mu_k}.}

\item{sigma}{A numeric vector of the scale parameter \eqn{\sigma_k}.}

\item{nu}{A numeric vector of the shape parameter \eqn{\nu_k}.}
}
\value{
\code{loglikelihood} returns the log-likelihood value.
}
\description{
Log-likelihood function.
}
